pyproject.toml
[project]
name = "ckanext-nhm"
version = "3.7.2"
description = "A CKAN extension for the Natural History Museum's Data Portal."
readme = "README.md"
requires-python = ">=3.6"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "Natural History Museum", email = "data@nhm.ac.uk" }
]
keywords = ["CKAN", "data", "nhm"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]
dependencies = [
    "ckantools",
    "cachetools>=4.2.4",
    "lxml>=4.4.2",
    "elasticsearch-dsl>=6.0.0,<7.0.0",
    "ckanext-ckanpackager==2.1.3",
    "ckanext-contact==2.0.1",
    "ckanext-doi==3.0.7",
    "ckanext-gallery==2.1.0",
    "ckanext-gbif==2.0.0",
    "ckanext-graph==2.0.1",
    "ckanext-ldap==3.1.2",
    "ckanext-query-dois==2.1.1",
    "ckanext-statistics==2.0.15",
    "ckanext-versioned-datastore==3.6.2",
    "ckanext-dcat @ git+https://github.com/ckan/ckanext-dcat@v1.3.0#egg=ckanext-dcat"
]

[project.optional-dependencies]
test = [
    "mock",
    "pytest>=4.6.5",
    "pytest-cov>=2.7.1",
    "coveralls"
]

[project.urls]
repository = "https://github.com/NaturalHistoryMuseum/ckanext-nhm"
changelog = "https://github.com/NaturalHistoryMuseum/ckanext-nhm/blob/main/CHANGELOG.md"

[project.entry-points."ckan.plugins"]
nhm = "ckanext.nhm.plugin:NHMPlugin"

[project.entry-points."ckan.rdf.plugins"]
nhm_dcat = "ckanext.nhm.dcat.profiles:NHMDCATProfile"


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[tool]
[tool.setuptools]
zip-safe = false

[tool.setuptools.packages.find]
exclude = ["tests", "docs"]

[tool.commitizen]
name = "cz_nhm"
version = "3.7.1"
tag_format = "v$version"
update_changelog_on_bump = true
changelog_incremental = false
version_files = [
    "package.json",
    "pyproject.toml:version"
]

[tool.black]
line-length = 88
skip_string_normalization = true

[tool.pylint]
max-line-length = 88
disable = ["C0114", "R0903"]

[tool.docformatter]
wrap-summaries = 88
wrap-descriptions = 88
pre-summary-newline = true
make-summary-multi-line = true
